add eip7702Support cli flag to toggle 7702 support
