+++ src/apis/web3-api.ts
@@ -1,4 +1,4 @@
-import { Web3API } from '../../types/index.js'
+import { Web3API } from '../types/index.js'
 
 export const createWeb3API = (version: string): Web3API => {
   return {
