+++ src/bundle/processor/bundle-processor.ts
@@ -12,9 +12,9 @@ import {
   ReputationManager,
 } from '../../types/index.js'
 import { packUserOps } from '../../utils/index.js'
-import { getUserOpHashes, selectBeneficiary } from './processor-helpers.js'
-import { findEntityToBlame, checkFatal } from '../bundle-helper.js'
-import { parseFailedOpRevert } from '../builder/builder-helpers.js'
+import { getUserOpHashes, selectBeneficiary } from './processor.helpers.js'
+import { findEntityToBlame, checkFatal } from '../bundle.helper.js'
+import { parseFailedOpRevert } from '../builder/builder.helpers.js'
 
 export type BundleProcessorConfig = {
   providerService: ProviderService
