+++ src/config/command.ts
@@ -92,6 +92,11 @@ export const getCmdOptionValues = (ops: CommandOptions): OptionValues => {
     )
     .option('--p2p', 'p2p mode enabled', false)
     .option('--findPeers', 'Search for peers when p2p is enabled.', false)
+    .option(
+      '--eip7702Support',
+      'On networks with EIP-7702 enabled, the eth_sendUserOperation method accepts an extra eip7702Auth parameter.',
+      true,
+    )
 
   const programOpts: OptionValues = program.parse(args).opts()
   return programOpts
