--- src/rpc/apis/index.ts
@@ -1,3 +1,0 @@
-export * from './Debug.js'
-export * from './Eth.js'
-export * from './Web3.js'
