+++ src/rpc/index.ts
@@ -1,3 +1,2 @@
 export * from './rpc-server.js'
 export * from './bundler-handler-registry.js'
-export * from './apis/index.js'
