+++ src/types/eip-7702.types.ts
@@ -1,0 +1,10 @@
+import type { BigNumberish } from 'ethers'
+
+export type EIP7702Authorization = {
+  chainId: BigNumberish
+  address: string
+  nonce: BigNumberish
+  yParity: BigNumberish
+  r: BigNumberish
+  s: BigNumberish
+}
