+++ src/types/express.d.ts
@@ -1,5 +1,7 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import type { Request } from 'express'
+import type { ValidatedJsonRpcRequest } from './rpc.types.js'
+import type { JsonRpcRequest } from './rpc.types.js'
 
 declare module 'express' {
   interface Request {
