+++ src/types/index.ts
@@ -12,4 +12,3 @@ export * from './bundler-collector-tracer.types.js'
 export * from './sim-types.js'
 export * from './api.types.js'
 export * from './error.types.js'
-export * from './handler.types.js'
